refactor: repository 패키지 domain으로 이동
- domain service를 만들다보니, repository <-> domain간 양방향이 발생하여 repository를 domain으로 이동하게됨!
