+++ src/main/java/kitchenpos/menu/domain/repository/MenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.repository;
+package kitchenpos.menu.domain.repository;
 
 import kitchenpos.menu.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
