+++ src/main/java/kitchenpos/menu/domain/repository/MenuRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.menu.repository;
+package kitchenpos.menu.domain.repository;
 
-import java.util.List;
 import kitchenpos.menu.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -8,5 +7,4 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface MenuRepository extends JpaRepository<Menu, Long> {
 
-    long countByIdIn(List<Long> ids);
 }
